Update Beet and Lepsen; clean up Beet config
Lepsen Core previously failed to build on Beet 1.35.1, but that has been
fixed so this commit upgrades to Lepsen Core 0.1.1 and Beet 1.35.1.
There should be no changes to generated data pack as a result of this,
though.

The main change is that Beet configuration has been moved to `beet.yaml`
instead of being shoved into `pyproject.toml`, so the pack description
and plugin configs are much cleaner.