Output data pack as both a folder and a zipfile
This simplifies the build/testing process as it guarantees that once the
folder has been tested, the zipfile can simply be directly uploaded to a
GitHub release without any manual changes or altered command invocation.
Instead of the data pack being output as a folder named "build", it will
be output as a subfolder and zipfile contained in the "build" directory.

Unfortunately to make this work without duplicated project config, we
have to `extend` a separate Beet config file, meaning our config will no
longer be named `beet.yaml` as is standard, so it is now named
`beet-project.yaml`. Fortunately, we avoid the confusing scenario where
we would have both a `beet.yaml` and `beet-project.yaml` by configuring
Beet in `pyproject.toml` to simply build both a folder and zipfile from
the configuration in `beet-project.yaml`, so the configuration is still
very straightfoward!